move filter section into separate view
